Patch by Ulrich Weigand:
A couple of tests fail simply because the code is not
safe on 64-bit big-endian platforms.

In MultiSource/Applications/viterbi, the read_dmatrix
routine uses scanf with a "%d" specifier to read into
a "size_t" variable -- this breaks when size_t is 64-bit
(may happen to work on little-endian, but never on
big-endian platforms).

In MultiSource/Benchmarks/MiBench/consumer-lame, the
DetermineByteOrder routine has an implicit assumption
that "long" is a 4-byte data type; again, it happens to
still work on a 64-bit little-endian system but breaks
on a big-endian system.

The committed patch fixes both problems, and both
tests now pass on PowerPC.  For viterbi, the patch
simply uses "int" instead of "size_t".  For consumer-lame,
the patch also just uses "int" instead of "long" -- this of
course still has an assumption that "int" is a 4-byte type,
but that's probably always true on systems where lame
runs in the first place.  (If not, we can find a more generic
way to determine endianness.  I just wanted to keep
changes to the benchmark sources as small as possible.)

llvm-svn: 165724

+++ MultiSource/Applications/viterbi/read_dmatrix.c
@@ -11,7 +11,7 @@ static const char *mybasename(const char *str) {
 size_t read_dmatrix(dvarray* out, const char* filename) {
     FILE* fid = fopen(filename,"rt");
     double value;
-    size_t i, j, width = 0, height = 0;
+    int i, j, width = 0, height = 0;
 
     printf("Opened file %s for matrix reading\n", mybasename(filename));
